Windows port: experimental quasar: MSVC narrowing, builtins, and always_inline fixes
- binary_ng: __builtin_unreachable() -> ttsl::unreachable() (device_operation.cpp,
  program_factory.cpp, binary_ng_utils.cpp); static_cast<uint32_t> for a
  CoreCoord-derived shard_shape braced init in device_operation.cpp.
- slice/device_operation.cpp: __attribute__((always_inline)) -> TT_FORCE_INLINE.
- transformer/sdpa_decode: __builtin_clz/__builtin_ctz -> tt::compiler::
  count_leading_zeros32/count_trailing_zeros32 from hostdevcommon/tt_compiler.h
  (device_operation.hpp, program_factory.cpp).
- matmul (dram-sharded and mcast_1d factories, sparse mcast_1d factory),
  padded_slice (rm/tile), pool_generic (multi-core factory, generic_pools),
  reshape_view (rm factory), reshard (nd copy-pages and generic factories):
  add static_cast<uint32_t>/<uint8_t> around size_t/int/DeviceAddr expressions
  that MSVC treats as narrowing inside aggregate and std::vector<uint32_t>
  braced-init-lists.

+++ ttnn/cpp/ttnn/operations/experimental/quasar/pool_generic/generic_pools.cpp
@@ -577,7 +577,10 @@ public:
         }
         return {
             {{input_slice_height_start, input_slice_width_start}, {input_slice_height_end, input_slice_width_end}},
-            {pad_top, pad_bottom, pad_left, pad_right},
+            {static_cast<uint32_t>(pad_top),
+             static_cast<uint32_t>(pad_bottom),
+             static_cast<uint32_t>(pad_left),
+             static_cast<uint32_t>(pad_right)},
             this_ceil_pad,
             output_slice_width};
     }
